+++ app/src/main/cpp/SineWavetable.cpp
@@ -13,7 +13,7 @@ void SineWavetable::generate()
 
     for (int i = 0; i < size; ++i)
     {
-        table[i] = sinf(currentAngle);
+        table[i] = sin(currentAngle);
         currentAngle += angleDelta;
     }
 }
