+++ app/src/main/cpp/WavetableOscillator.h
@@ -12,27 +12,25 @@
 class WavetableOscillator
 {
 protected:
-    static float constexpr kPI = M_PI;
-    static float constexpr kTwoPi = kPI * 2;
+    static double constexpr kPI = M_PI;
+    static double constexpr kTwoPi = kPI * 2.0;
     const Wavetable* wavetable;
     int tableSize;
-    //float currentIndex = 0.0f;
-    float accumulatedPhase = 0.0f;
-    //float tableDelta = 0.0f;
-    float phaseDelta;
-    float frequency = 0.0f;
-    float gain = 0.0f;
-    float currentSample = 0.0f;
+    double accumulatedPhase = 0.0;
+    double phaseDelta;
+    double frequency = 0.0;
+    double gain = 0.0;
+    double currentSample = 0.0;
 public:
     WavetableOscillator(const Wavetable* wavetableToUse);
     virtual ~WavetableOscillator();
 
-    void setFrequency(float freq, float sampleRate);
-    float getFrequency() const;
+    void setFrequency(double freq, double sampleRate);
+    double getFrequency() const;
     void setWavetable(const Wavetable* newWavetable);
-    void setGain(float newGain);
-    virtual inline float getNextSample() noexcept = 0;
-    float getCurrentSample() const { return currentSample; };
+    void setGain(double newGain);
+    virtual inline double getNextSample() noexcept = 0;
+    double getCurrentSample() const { return currentSample; };
 };
